Remove specialized components from utils and add to corresponding folder. Move specific utils to specific folders, and reorganize calendar and form.
