+++ src/components/general/Form/index.js
@@ -1,0 +1,1 @@
+export * from './Form';
