+++ src/components/general/index.js
@@ -1,4 +1,5 @@
-export * from './calendar';
+export * from './Calendar';
+export * from './Form';
 export * from './Alert';
 export * from './Animation';
 export * from './Avatar';
