+++ src/components/specialized/EventPanel.js
