+++ src/components/specialized/index.js
@@ -1,0 +1,3 @@
+export * from './EventPanel';
+export * from './MemberPanel';
+export * from './DefaultAgenda';
