--- src/utils/actions.js
@@ -1,3 +1,0 @@
-export const createActionTypes = actionTypes => actionTypes
-	.map(type => ({ [type]: `module/${type}`	}))
-	.reduce((types, type) => ({ ...types, ...type }), {});
